Record node canvas shape IDs
This will let us track when a node is clicked by knowing
which canvas shapes belong to which node.

Each node is responsible for setting his/her own canvas
ids, which are returned by the canvas create_XXX
methods. This is different each frame, so it must be done
in the the _draw function as the shapes are created.